Cleaned demo projects
