+++ Demos/GettingStarted/GettingStarted.SqlServer.Autofac/GettingStarted.SqlServer.Autofac.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp2.2</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\SqlRepo.SqlServer.Autofac\SqlRepo.SqlServer.Autofac.csproj" />
+    <ProjectReference Include="..\GettingStarted.SqlServer.Shared\GettingStarted.SqlServer.Shared.csproj" />
+  </ItemGroup>
+
+</Project>
