+++ Demos/GettingStarted/GettingStarted.SqlServer.Ninject/GettingStarted.SqlServer.Ninject.csproj
@@ -1,0 +1,18 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp2.2</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Ninject" Version="3.3.4" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\SqlRepo.SqlServer.NInject\SqlRepo.SqlServer.Ninject.csproj" />
+    <ProjectReference Include="..\GettingStarted.SqlServer.Shared\GettingStarted.SqlServer.Shared.csproj" />
+  </ItemGroup>
+
+
+</Project>
