--- Demos/GettingStarted/GettingStarted/IGettingStarted.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace GettingStartedIoC
-{
-    public interface IGettingStarted
-    {
-        void DoIt();
-    }
-}
