--- Demos/GettingStarted/GettingStarted/packages.config
@@ -1,9 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Autofac" version="4.8.1" targetFramework="net47" />
-  <package id="System.Configuration.ConfigurationManager" version="4.5.0" targetFramework="net47" />
-  <package id="System.Data.SqlClient" version="4.5.1" targetFramework="net47" />
-  <package id="System.Security.AccessControl" version="4.5.0" targetFramework="net47" />
-  <package id="System.Security.Permissions" version="4.5.0" targetFramework="net47" />
-  <package id="System.Security.Principal.Windows" version="4.5.0" targetFramework="net47" />
-</packages>
