--- Demos/GettingStarted/GettingStartedStatic/App.config
@@ -1,13 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-
-<configuration>
-  <connectionStrings>
-    <clear />
-    <add name="GettingStartedDbCon"
-         connectionString="Data Source=(LocalDB)\MSSQLLocalDB;AttachDbFilename=|DataDirectory|GettingStarted.mdf;Integrated Security=True;Connect Timeout=30" />
-  </connectionStrings>
-  <startup>
-    <supportedRuntime version="v4.0"
-                      sku=".NETFramework,Version=v4.7" />
-  </startup>
-</configuration>
