--- GettingStarted.SqlServer.Ninject/GettingStarted.SqlServer.Ninject.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Ninject" Version="3.3.4" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Demos\GettingStarted\GettingStarted.SqlServer.Shared\GettingStarted.SqlServer.Shared.csproj" />
-    <ProjectReference Include="..\SqlRepo.SqlServer.NInject\SqlRepo.SqlServer.Ninject.csproj" />
-  </ItemGroup>
-
-</Project>
