--- GettingStarted.SqlServer.ServiceCollection/GettingStarted.SqlServer.ServiceCollection.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="2.2.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Demos\GettingStarted\GettingStarted.SqlServer.Shared\GettingStarted.SqlServer.Shared.csproj" />
-    <ProjectReference Include="..\SqlRepo.SqlServer.ServiceCollection\SqlRepo.SqlServer.ServiceCollection.csproj" />
-  </ItemGroup>
-
-</Project>
