+++ SqlRepo.SqlServer/Abstractions/ISelectClauseBuilder.cs
@@ -18,7 +18,6 @@ namespace SqlRepo.SqlServer.Abstractions
 
         ISelectClauseBuilder CountAll();
 
-        [Obsolete]
         ISelectClauseBuilder For<TEntity>(TEntity entity,
             string alias = null,
             string tableName = null,
