+++ SqlRepo.sln
@@ -19,6 +19,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 	ProjectSection(SolutionItems) = preProject
 		appveyor.yml = appveyor.yml
 		BumpVersion.ps1 = BumpVersion.ps1
+		PatchProjectVersions.ps1 = PatchProjectVersions.ps1
 		Version.txt = Version.txt
 	EndProjectSection
 EndProject
@@ -36,15 +37,15 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Demos", "Demos", "{342408F0
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "GettingStarted", "GettingStarted", "{D58D61C5-3D97-45CF-A3B0-3968E3407EFF}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GettingStarted.SqlServer.Autofac", "Demos\GettingStarted\GettingStarted.SqlServer.Ioc\GettingStarted.SqlServer.Autofac.csproj", "{89A55EA8-92D0-4B00-BBC3-7719D51AC296}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GettingStarted.SqlServer.Shared", "Demos\GettingStarted\GettingStarted.SqlServer.Shared\GettingStarted.SqlServer.Shared.csproj", "{D5F56892-D523-401B-9595-654B9653B4BD}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GettingStarted.SqlServer.Shared", "Demos\GettingStarted\GettingStarted.SqlServer.Shared\GettingStarted.SqlServer.Shared.csproj", "{D373733B-2B6C-4244-B0C4-8161818D6923}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GettingStarted.SqlServer.Static", "Demos\GettingStarted\GettingStarted.SqlServer.Static\GettingStarted.SqlServer.Static.csproj", "{DF95EB23-303B-4053-AD75-C2B9278ED9BE}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GettingStarted.SqlServer.Static", "Demos\GettingStarted\GettingStarted.SqlServer.Static\GettingStarted.SqlServer.Static.csproj", "{588797CE-7476-4D6B-8B8B-0974D7C6CFD7}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GettingStarted.SqlServer.ServiceCollection", "Demos\GettingStarted\GettingStarted.SqlServer.ServiceCollection\GettingStarted.SqlServer.ServiceCollection.csproj", "{8CDC27A2-81CC-4814-8E46-B13B4551E95E}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GettingStarted.SqlServer.ServiceCollection", "GettingStarted.SqlServer.ServiceCollection\GettingStarted.SqlServer.ServiceCollection.csproj", "{882D0EB0-7704-4851-B176-DA12551E9D33}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GettingStarted.SqlServer.Ninject", "Demos\GettingStarted\GettingStarted.SqlServer.Ninject\GettingStarted.SqlServer.Ninject.csproj", "{B349E7B1-8352-45D3-AC22-C3A699C862F3}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GettingStarted.SqlServer.Ninject", "GettingStarted.SqlServer.Ninject\GettingStarted.SqlServer.Ninject.csproj", "{1829862F-1CC9-4216-B428-F3305E79C2F3}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GettingStarted.SqlServer.Autofac", "Demos\GettingStarted\GettingStarted.SqlServer.Autofac\GettingStarted.SqlServer.Autofac.csproj", "{63E3EB90-9783-49B6-A762-5EA2776E0570}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -96,37 +97,37 @@ Global
 		{C090C6B6-CA18-4708-88C5-15AB9576C33C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C090C6B6-CA18-4708-88C5-15AB9576C33C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C090C6B6-CA18-4708-88C5-15AB9576C33C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{89A55EA8-92D0-4B00-BBC3-7719D51AC296}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{89A55EA8-92D0-4B00-BBC3-7719D51AC296}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{89A55EA8-92D0-4B00-BBC3-7719D51AC296}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{89A55EA8-92D0-4B00-BBC3-7719D51AC296}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D373733B-2B6C-4244-B0C4-8161818D6923}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D373733B-2B6C-4244-B0C4-8161818D6923}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D373733B-2B6C-4244-B0C4-8161818D6923}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D373733B-2B6C-4244-B0C4-8161818D6923}.Release|Any CPU.Build.0 = Release|Any CPU
-		{588797CE-7476-4D6B-8B8B-0974D7C6CFD7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{588797CE-7476-4D6B-8B8B-0974D7C6CFD7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{588797CE-7476-4D6B-8B8B-0974D7C6CFD7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{588797CE-7476-4D6B-8B8B-0974D7C6CFD7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{882D0EB0-7704-4851-B176-DA12551E9D33}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{882D0EB0-7704-4851-B176-DA12551E9D33}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{882D0EB0-7704-4851-B176-DA12551E9D33}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{882D0EB0-7704-4851-B176-DA12551E9D33}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1829862F-1CC9-4216-B428-F3305E79C2F3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1829862F-1CC9-4216-B428-F3305E79C2F3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1829862F-1CC9-4216-B428-F3305E79C2F3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1829862F-1CC9-4216-B428-F3305E79C2F3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D5F56892-D523-401B-9595-654B9653B4BD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D5F56892-D523-401B-9595-654B9653B4BD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D5F56892-D523-401B-9595-654B9653B4BD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D5F56892-D523-401B-9595-654B9653B4BD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DF95EB23-303B-4053-AD75-C2B9278ED9BE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DF95EB23-303B-4053-AD75-C2B9278ED9BE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DF95EB23-303B-4053-AD75-C2B9278ED9BE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DF95EB23-303B-4053-AD75-C2B9278ED9BE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8CDC27A2-81CC-4814-8E46-B13B4551E95E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8CDC27A2-81CC-4814-8E46-B13B4551E95E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8CDC27A2-81CC-4814-8E46-B13B4551E95E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8CDC27A2-81CC-4814-8E46-B13B4551E95E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B349E7B1-8352-45D3-AC22-C3A699C862F3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B349E7B1-8352-45D3-AC22-C3A699C862F3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B349E7B1-8352-45D3-AC22-C3A699C862F3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B349E7B1-8352-45D3-AC22-C3A699C862F3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{63E3EB90-9783-49B6-A762-5EA2776E0570}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{63E3EB90-9783-49B6-A762-5EA2776E0570}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{63E3EB90-9783-49B6-A762-5EA2776E0570}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{63E3EB90-9783-49B6-A762-5EA2776E0570}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{D58D61C5-3D97-45CF-A3B0-3968E3407EFF} = {342408F0-6D7A-44D2-A002-D1730A7DBB81}
-		{89A55EA8-92D0-4B00-BBC3-7719D51AC296} = {D58D61C5-3D97-45CF-A3B0-3968E3407EFF}
-		{D373733B-2B6C-4244-B0C4-8161818D6923} = {D58D61C5-3D97-45CF-A3B0-3968E3407EFF}
-		{588797CE-7476-4D6B-8B8B-0974D7C6CFD7} = {D58D61C5-3D97-45CF-A3B0-3968E3407EFF}
-		{882D0EB0-7704-4851-B176-DA12551E9D33} = {D58D61C5-3D97-45CF-A3B0-3968E3407EFF}
-		{1829862F-1CC9-4216-B428-F3305E79C2F3} = {D58D61C5-3D97-45CF-A3B0-3968E3407EFF}
+		{D5F56892-D523-401B-9595-654B9653B4BD} = {D58D61C5-3D97-45CF-A3B0-3968E3407EFF}
+		{DF95EB23-303B-4053-AD75-C2B9278ED9BE} = {D58D61C5-3D97-45CF-A3B0-3968E3407EFF}
+		{8CDC27A2-81CC-4814-8E46-B13B4551E95E} = {D58D61C5-3D97-45CF-A3B0-3968E3407EFF}
+		{B349E7B1-8352-45D3-AC22-C3A699C862F3} = {D58D61C5-3D97-45CF-A3B0-3968E3407EFF}
+		{63E3EB90-9783-49B6-A762-5EA2776E0570} = {D58D61C5-3D97-45CF-A3B0-3968E3407EFF}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {788D8B76-6423-427D-8CA2-FD6C6E977803}
